Move file system stuff into own dir
