+++ src/file_system_interaction/asset_loading.rs
