+++ src/lib.rs
@@ -8,29 +8,29 @@ mod audio;
 mod condition;
 mod dev;
 mod dialog;
+mod file_system_interaction;
 mod interactions_ui;
-mod loading;
 mod map;
 mod math;
 mod menu;
 pub mod mesh_util;
 mod movement_gameplay;
 mod physics;
-mod saving;
 #[cfg(feature = "editor")]
 mod scene_editor;
 mod shader;
 mod spawning;
 mod trait_extension;
-mod world_serialization;
 
 use crate::actions::ActionsPlugin;
 use crate::audio::InternalAudioPlugin;
 use crate::condition::ConditionPlugin;
 use crate::dev::DevPlugin;
 use crate::dialog::DialogPlugin;
+use crate::file_system_interaction::asset_loading::LoadingPlugin;
+use crate::file_system_interaction::game_serialization::SavingPlugin;
+use crate::file_system_interaction::level_serialization::WorldSerializationPlugin;
 use crate::interactions_ui::InteractionsUi;
-use crate::loading::LoadingPlugin;
 use crate::map::MapPlugin;
 use crate::menu::MenuPlugin;
 use crate::movement_gameplay::camera::CameraPlugin;
@@ -38,10 +38,8 @@ use crate::movement_gameplay::general_movement::MovementPlugin;
 use crate::movement_gameplay::navigation::NavigationPlugin;
 use crate::movement_gameplay::player::PlayerPlugin;
 use crate::physics::PhysicsPlugin;
-use crate::saving::SavingPlugin;
 use crate::shader::ShaderPlugin;
 use crate::spawning::SpawningPlugin;
-use crate::world_serialization::WorldSerializationPlugin;
 use bevy::app::App;
 use bevy::prelude::*;
