+++ src/scene_editor.rs
@@ -1,10 +1,10 @@
 use crate::actions::{Actions, ActionsFrozen};
-use crate::saving::{GameLoadRequest, GameSaveRequest};
+use crate::file_system_interaction::game_serialization::{GameLoadRequest, GameSaveRequest};
+use crate::file_system_interaction::level_serialization::{WorldLoadRequest, WorldSaveRequest};
 use crate::spawning::{
     DelayedSpawnEvent, DuplicationEvent, GameObject, ParentChangeEvent,
     SpawnEvent as SpawnRequestEvent,
 };
-use crate::world_serialization::{WorldLoadRequest, WorldSaveRequest};
 use crate::GameState;
 use bevy::prelude::*;
 use bevy_egui::egui::ScrollArea;
